GP: Don't treat display of publication as a query
When displaying a publication in Google Patents, the url includes the
query string that you used to find that publication.  In previous
versions, the history generator found the url, parsed it, and displayed
it in the history even though it didn't correspond to an actual query.
The new version ignores those.